Update flipper parser: exit codes of premature exit found from output file
If the output file showed that the calculation ended prematurely, return
the corresponding exit codes (e.g. out of walltime).
This uses the `validate_premature_exit` method from the PwParser.

Notice that, in contrast to the PwCalculation, some of the exit codes of
the FlipperCalculation may actually be recovered, if a trajectory was
produced.